Implemented basic path browsing! Yipeeeee! Users can now select a drive from the dropdown, and the contents of that drive will be displayed in the listbox below.
Selecting a folder from the listbox will update the listbox to show the
contents of that folder.

We also added basic path navigation (e.g. "C:\Users\Peter\Documents") so that
users can see their current path and navigate back up the directory tree.

Also we added a clear button to reset the path and listbox to the initial state.
That way users can also select a different drive if they want to.