Upgraded a lot of stuff, included threading We now use queues implementations to manage subdirs indexing and UI rendering on separate threads more efficiently. This results in overall better performance when loading large folders.
We also added a progressbar to see the indexing progress and some
more info like your current version of nanoFiler, and the current path.

We plan on refactoring the file/folder logic to OOP (Object Oriented
Programming) in order to make the code more maintainable and easier to
upgrade in the future.