We did it! We refactored to OOP! Yay!
Our file explorer is still very raw, turns out there was A LOT of
unused code in there, but we're still working on that. We're aware
that nanoFiler is still not fully optimised, so we're looking forward
to fixing that.

We're also working on how to display video files, and doing basic
operations like creating new files, editing text files and deleting all
sorts of other files.